Add more robust tests for set_option_defaults
