+++ setup.cfg
@@ -64,6 +64,8 @@ testing =
 	sphinx
 	jaraco.path>=3.2.0
 
+	build
+
 docs =
 	# upstream
 	sphinx
